+++ tests/test_renderer.py
@@ -453,7 +453,7 @@ See `https://example.com`_.
 
     def test_disable_inline_math(self):
         src = "this is `$E = mc^2$` inline math."
-        out = self.conv(src, disable_inline_math=True)
+        out = self.conv(src, inline_math=None)
         self.assertEqual(out, "\nthis is ``$E = mc^2$`` inline math.\n")
 
     def test_inline_html(self):